+++ components/progress/BarProgress.js
@@ -1,56 +1,87 @@
 import React from 'react'
 import Classnames from 'classnames'
 
+const defaultUnit = 'px'
+const isNumeric = (n) => !isNaN(parseFloat(n)) && isFinite(n)
+
 export default class BarProgress extends React.Component {
   textRef = React.createRef()
+  barRef = React.createRef()
+
   state = { insidePlacement: 'right' }
 
-  componentDidMount () {
+  componentDidMount() {
     if (this.props.placement === 'inside') {
-      if (
-        this.textRef.current &&
-        this.textRef.current.clientWidth >= this.getWidth() - ((this.getWidth() * this.props.percent) / 100 + 5)
-      ) {
-        this.setState({ insidePlacement: 'left' })
-      } else {
-        this.setState({ insidePlacement: 'right' })
-      }
+      const { current: barElm } = this.barRef
+      const { current: textElm } = this.textRef
+      const barClientWidth = barElm ? barElm.clientWidth : 0
+
+      const isOver =
+        textElm && textElm.clientWidth >= barClientWidth - ((barClientWidth * this.props.percent) / 100 + 5)
+      const insidePlacement = isOver ? 'left' : 'right'
+
+      this.setState({
+        insidePlacement
+      })
     }
-  }
-  getWidth = () => {
-    const { width, size } = this.props
-    if (!width || width <= 0) {
-      return size === 'large' ? 480 : 160
-    }
-    return width
-  }
-  getHeight = () => {
-    const { size, height } = this.props
-    if (!height || height <= 0) {
-      return size === 'large' ? 8 : size === 'default' ? 6 : 2
-    }
-    return height
   }
 
-  render () {
-    let prefix = 'hi-progress'
-    const { percent: percentNum, content, type, showInfo, placement, tooltip = null, active, text, withOutText, status  } = this.props
+  get width() {
+    const { width, size } = this.props
+    const num = parseInt(width) || 0
+
+    if (num <= 0) {
+      const defaultNum = size === 'large' ? 480 : 160
+      return defaultNum + defaultUnit
+    }
+
+    return isNumeric(width) ? width + defaultUnit : width
+  }
+
+  get height() {
+    const { height, size } = this.props
+    const num = parseInt(height) || 0
+
+    if (num <= 0) {
+      const defaultNum = size === 'large' ? 8 : size === 'default' ? 6 : 2
+      return defaultNum + defaultUnit
+    }
+
+    return isNumeric(height) ? height + defaultUnit : height
+  }
+
+  render() {
+    const prefix = 'hi-progress'
+    const {
+      percent: percentNum,
+      content,
+      type,
+      showInfo,
+      placement,
+      tooltip = null,
+      active,
+      text,
+      withOutText,
+      status
+    } = this.props
+    const { width, height } = this
 
     const _content = typeof content !== 'undefined' ? content : text // // api 兼容 1.x 为 text 2.x 改为 content
     const _showInfo = typeof showInfo !== 'undefined' ? showInfo : withOutText // // api 兼容 1.x 为 withOutText 2.x 改为 showInfo
     const _type = type || status
+    const { current: barElm } = this.barRef
 
     const percent = percentNum > 0 ? percentNum : 0
     return (
       <div>
-        <div className={`${prefix}__inner`} style={{ width: this.getWidth() + 'px', height: this.getHeight() + 'px' }}>
+        <div ref={this.barRef} className={`${prefix}__inner`} style={{ width, height }}>
           <div
             className={Classnames(`${prefix}__bar ${prefix}__bar--${_type}`, {
               [`${prefix}__bar--active`]: active
             })}
             style={{ width: `${percent}%` }}
           >
-            {_showInfo && placement === 'inside' && this.getHeight() >= 14 && (
+            {_showInfo && placement === 'inside' && (!barElm || barElm.clientHeight >= 14) && (
               <div ref={this.textRef} className={`${prefix}__text--inside inside--${this.state.insidePlacement}`}>
                 {_content || `${percent}%`}
               </div>